Move `Transformer` from `pipe` to `pipeline.transformer`
