+++ curated_transformers/tests/pipeline/toy-en-corpus.spacy
