+++ setup.cfg
@@ -20,7 +20,7 @@ install_requires =
 
 [options.entry_points]
 spacy_factories =
-    curated_transformer = curated_transformers.pipe:make_transformer
+    curated_transformer = curated_transformers.pipeline.transformer:make_transformer
 
 spacy_architectures =
     curated-transformers.BertTransformer.v1 = curated_transformers.models:build_bert_transformer_model_v1
